+++ tests/NATS.Client.Core2.Tests/ProtocolTest.cs
@@ -1,9 +1,7 @@
 using System.Buffers;
 using System.Text;
 using NATS.Client.Core2.Tests;
-#if !NET6_0_OR_GREATER
-using NATS.Client.Core.Internal.NetStandardExtensions;
-#endif
+using NATS.Client.Core2.Tests.ExtraUtils.FrameworkPolyfillExtensions;
 
 namespace NATS.Client.Core.Tests;
 
@@ -489,7 +487,7 @@ public class ProtocolTest
 
         protected override ValueTask ReceiveInternalAsync(string subject, string? replyTo, ReadOnlySequence<byte>? headersBuffer, ReadOnlySequence<byte> payloadBuffer)
         {
-            _callback(int.Parse(Encoding.UTF8.GetString(payloadBuffer)));
+            _callback(int.Parse(Encoding.UTF8.GetString(payloadBuffer.ToArray())));
             DecrementMaxMsgs();
             return default;
         }
