+++ tests/NATS.Client.TestUtilities2/NATS.Client.TestUtilities2.csproj
@@ -6,4 +6,8 @@
         <IsPackable>false</IsPackable>
     </PropertyGroup>
 
+    <ItemGroup>
+      <ProjectReference Include="..\..\src\NATS.Net\NATS.Net.csproj" />
+    </ItemGroup>
+
 </Project>
